--В связи с изменением работы программы, были удалены классы InMemoryFilmStorage, InMemoryUserStorage, LikesAmountComparator и BasicService; --Теперь информация о фильмах, пользователях и их взаимодействии хранится в базах данных, для чего были изменены интерфейсы FilmStorage и UserStorage, а также добавлены новые: GenreStorage и RatingStorage; --За реализацию указанных интерфейсов отвечают классы: FilmRepository, UserRepository, GenreRepository и MpaRepository, хранящиеся в пакете dal. Их общие функции собраны в родительском для них классе BaseRepository; --Был добавлен ряд вспомогательных классов в подпакет mappers внутри пакета dal для преобразования данных из базы; --В связи с расширением функций программы был добавлен ряд обрабатываемых исключений в пакет exception; --Для проверки новых методов был добавлен ряд тестов в соответствующий пакет;

+++ src/main/java/ru/yandex/practicum/filmorate/exception/FilmIdException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.exception;
 
 public class FilmIdException extends RuntimeException {
-    public FilmIdException(int id) {
+    public FilmIdException(long id) {
         super(String.format("Фильм с id %d не найден", id));
     }
 }
